Reorganize mock-related code
